chore(layout and doc): - Module local clients are moved under ./domain folders - docs are updated
