+++ examples/modkit/users_info/users_info/src/domain/local_client/mod.rs
