+++ examples/modkit/users_info/users_info/src/domain/mod.rs
@@ -1,5 +1,6 @@
 pub mod error;
 pub mod events;
+pub mod local_client;
 pub mod ports;
 pub mod repos;
 pub mod service;
