+++ examples/modkit/users_info/users_info/src/infra/mod.rs
@@ -1,3 +1,2 @@
 pub mod audit;
-pub mod local_client;
 pub mod storage;
