+++ examples/plugin-modules/tenant_resolver/tenant_resolver-gw/src/domain/local_client.rs
