+++ examples/plugin-modules/tenant_resolver/tenant_resolver-gw/src/domain/mod.rs
@@ -1,2 +1,4 @@
 pub mod error;
+#[doc(hidden)]
+pub mod local_client;
 pub mod service;
