+++ examples/plugin-modules/tenant_resolver/tenant_resolver-gw/src/lib.rs
@@ -14,5 +14,3 @@ pub mod api;
 pub mod config;
 #[doc(hidden)]
 pub mod domain;
-#[doc(hidden)]
-pub mod local_client;
